+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessCSSFilesTask.java
@@ -62,15 +62,16 @@ public class ProcessCSSFilesTask extends ProcessFilesTask {
      * @param outputFilename the output file name
      * @param engine minify processor engine selected
      * @param yuiConfig YUI Compressor configuration
+     * @param newLineBetweenFiles Add a new line between every files
      * @throws FileNotFoundException when the given source file does not exist
      */
     public ProcessCSSFilesTask(Log log, boolean verbose, Integer bufferSize, String charset, String suffix,
             boolean nosuffix, boolean skipMerge, boolean skipMinify, String webappSourceDir, String webappTargetDir,
             String inputDir, List<String> sourceFiles, List<String> sourceIncludes, List<String> sourceExcludes,
-            String outputDir, String outputFilename, Engine engine, YuiConfig yuiConfig) throws FileNotFoundException {
+            String outputDir, String outputFilename, Engine engine, YuiConfig yuiConfig, boolean newLineBetweenFiles) throws FileNotFoundException {
         super(log, verbose, bufferSize, charset, suffix, nosuffix, skipMerge, skipMinify, webappSourceDir,
                 webappTargetDir, inputDir, sourceFiles, sourceIncludes, sourceExcludes, outputDir, outputFilename,
-                engine, yuiConfig);
+                engine, yuiConfig, newLineBetweenFiles);
     }
 
     /**